requirements.txt
# DocSynthesis-V1 Core Dependencies

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# NLP and ML
sentencepiece>=0.1.99
tokenizers>=0.14.0
numpy>=1.24.0
scipy>=1.11.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Monitoring and Logging
loguru>=0.7.0
prometheus-client>=0.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Optional: for PDF processing
PyPDF2>=3.0.0
pdf2image>=1.16.0

# Optional: for document conversion
python-docx>=1.0.0

